Create multiplication.py
This is my side project.

Why am i made it, because lots of peoples made it... Also basically code.... said it from of them...

yes I think it made for hard to me... 

But! I made it!!! :)

If you have a any ideas of my code, comments please!

I will develope to my code!!!

Thank you!